requirements.txt
Django>=2.1.3,<2.20
djangorestframework>=3.9.4,<3.10.0
flake8>=3.6.0,<3.8.0
autopep8>=1.4.4,<1.5.0
coveralls>=1.8.0,<1.9.0
coverage>=4.5.3,<4.6.0
psycopg2>=2.7.5,<2.9.0
